BJS2-26621 into master + refactored test packages
