+++ src/main/java/school/faang/user_service/mapper/recommendation/RecommendationRequestMapper.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.mapper.recomendation;
+package school.faang.user_service.mapper.recommendation;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
