+++ src/test/java/school/faang/user_service/controller/recommendation/RecommendationRequestControllerTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.controller;
+package school.faang.user_service.controller.recommendation;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,12 +7,11 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import school.faang.user_service.controller.recommendation.RecommendationRequestController;
 import school.faang.user_service.dto.recommendation.RecommendationRequestDto;
 import school.faang.user_service.dto.recommendation.RejectionDto;
 import school.faang.user_service.dto.recommendation.RequestFilterDto;
 import school.faang.user_service.exception.DataValidationException;
-import school.faang.user_service.service.recomendation.RecommendationRequestService;
+import school.faang.user_service.service.recommendation.RecommendationRequestService;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
