+++ src/test/java/school/faang/user_service/controller/subscription/SubscriptionControllerTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.contoller;
+package school.faang.user_service.controller.subscription;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -8,7 +8,6 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import school.faang.user_service.controller.subscription.SubscriptionController;
 import school.faang.user_service.service.subscription.SubscriptionService;
 import school.faang.user_service.util.TestDataFactory;
 import school.faang.user_service.validator.skill.SubscriptionValidator;
